temporarily disable memory cache saving to database

+++ code/web/sys/MemoryCache/Memcache.php
@@ -3,12 +3,13 @@
 require_once ROOT_DIR . '/sys/MemoryCache/CachedValue.php';
 class Memcache
 {
+	private $enableDbCache = false;
     private $vars = array();
 
     public function get($name){
     	if (array_key_exists($name, $this->vars)) {
             return $this->vars[$name];
-        }else{
+        }elseif($this->enableDbCache){
     		try {
 			    $cachedValue = new CachedValue();
 			    $cachedValue->cacheKey = $name;
@@ -27,45 +28,49 @@ class Memcache
         return false;
     }
 
-    public function set($name, $value, $flag, $timeout) {
-        $this->vars[$name] = $value;
-        $valueToCache = serialize($value);
-        if (strlen($valueToCache) <= 1024 && strlen($name) < 200 ){
-	        try {
-		        $cachedValue = new CachedValue();
-		        $cachedValue->cacheKey = $name;
-		        $isNew = true;
-		        if ($cachedValue->find(true)){
-			        $isNew = false;
-		        }
-		        $cachedValue->value = $valueToCache;
-		        if ($timeout == 0){
-			        $cachedValue->expirationTime = 0;
-		        }else{
-			        $cachedValue->expirationTime = time() + $timeout;
-		        }
+    public function set($name, $value, $flag, $timeout)
+    {
+	    $this->vars[$name] = $value;
+	    if ($this->enableDbCache) {
+		    $valueToCache = serialize($value);
+		    if (strlen($valueToCache) <= 1024 && strlen($name) < 200) {
+			    try {
+				    $cachedValue = new CachedValue();
+				    $cachedValue->cacheKey = $name;
+				    $isNew = true;
+				    if ($cachedValue->find(true)) {
+					    $isNew = false;
+				    }
+				    $cachedValue->value = $valueToCache;
+				    if ($timeout == 0) {
+					    $cachedValue->expirationTime = 0;
+				    } else {
+					    $cachedValue->expirationTime = time() + $timeout;
+				    }
 
-		        if ($isNew){
-			        $cachedValue->insert();
-		        }else{
-			        $cachedValue->update();
-		        }
-	        }catch(Exception $e){
-		        //Table has not been created ignore
-	        }
-        }
-
+				    if ($isNew) {
+					    $cachedValue->insert();
+				    } else {
+					    $cachedValue->update();
+				    }
+			    } catch (Exception $e) {
+				    //Table has not been created ignore
+			    }
+		    }
+	    }
         return true;
     }
 
     public function delete($name){
         unset($this->vars[$name]);
-        try{
-		    $cachedValue = new CachedValue();
-		    $cachedValue->cacheKey = $name;
-		    $cachedValue->delete(true);
-        }catch(Exception $e){
-	        //Table has not been created ignore
-        }
+	    if($this->enableDbCache) {
+		    try {
+			    $cachedValue = new CachedValue();
+			    $cachedValue->cacheKey = $name;
+			    $cachedValue->delete(true);
+		    } catch (Exception $e) {
+			    //Table has not been created ignore
+		    }
+	    }
     }
 }
